Add dark class by default
Since it takes time for the page to hydrate and read the user's theme
preference from localStorage, there can be a small flash of the light
theme before it switches to the dark one. I figure it's better to start
dark and switch to light instead of flashing a bright light in the dark.